Routes

+++ src/usecases/dentists.js
@@ -21,14 +21,11 @@ async function createDentist(newDentist){
     try{
         const {email, password} = newDentist
         let emailExist = await Dentists.findOne({email: email})
-        if(emailExist){
-            throw new Error('email already in use, recover password or use ana nother email');
-        } else {
-            let encryptedPassword= await Bcrypt.hash(password);
-            return Dentist.create({...newDentist, password: encryptedPassword})
-        }
-    }
-    catch(error){console.log(error.message)}
+        if(emailExist) throw new Error('email already in use, recover password or use ana nother email');
+        
+        let encryptedPassword= await Bcrypt.hash(password);
+        return Dentists.create({...newDentist, password: encryptedPassword});
+    } catch(error){console.log(error.message)}
 }
 
 // .: Verifying existing dentist email for registration
@@ -38,12 +35,22 @@ async function createDentist(newDentist){
 
 // .: patch dentist
 async function updateDentist(id, newDentistData){
-    return Dentists.findByIdAndUpdate(id, newDentistData, {new: true})
+    try{
+        let idExist = Dentists.findById(id)
+        if(!idExist) throw new Error("Dentist doesn't exist");
+        return Dentists.findByIdAndUpdate(id, newDentistData, {new: true})
+    }
+    catch(error){console.log(error.message)}
 }
 
 // .: delete dentist
-async function deleteDentist(id){
-    return Dentists.findByIdAndDelete(id)
+async function deleteDentist(id) {
+    try {
+        let idExist = Dentists.findById(id)
+        if(!idExist) throw new Error("Dentist doesn't exist");
+        return Dentists.findByIdAndDelete(id)
+    }
+    catch(error){console.log(error.message)}
 }
 
 module.exports = {
